Fade in amimation hero section
