+++ src/app/globals.css
@@ -368,3 +368,20 @@ button .text-transparent {
   }
 }
 
+
+
+
+
+.fade-in-section {
+  opacity: 1;
+  transform: translateY(40vh);
+  visibility: 1;
+  transition: opacity 2s ease-out, transform 1s ease-out,
+    visibility 1200ms ease-out;
+  will-change: opacity, transform, visibility;
+}
+.fade-in-section.is-visible {
+  opacity: 1;
+  transform: none;
+  visibility: visible;
+}
