+++ src/app/page.js
@@ -1,7 +1,10 @@
+"use client";
+
 import Hero from "@/components/home/hero";
 import Navbar from "@/components/navbar";
 // import Footer from "@/components/footer";
 import Agree from "@/components/home/agree";
+import FadeInSection from "@/components/fadeInSection";
 
 export default function Home() {
   return (
@@ -10,7 +13,9 @@ export default function Home() {
         <div className="mb-10">
           <Navbar />
         </div>
-        <Hero />
+        <FadeInSection>
+          <Hero />
+        </FadeInSection>
         <Agree />
       </main>
     </div>
