--- src/components/home/style.css
@@ -1,13 +1,0 @@
-.fade-in-section {
-  opacity: 0;
-  transform: translateY(40vh);
-  visibility: hidden;
-  transition: opacity 2s ease-out, transform 1s ease-out,
-    visibility 1200ms ease-out;
-  will-change: opacity, transform, visibility;
-}
-.fade-in-section.is-visible {
-  opacity: 1;
-  transform: none;
-  visibility: visible;
-}
